Scrambler v0.9
Takes an input sentence and scrambles all of the letters in each word
(Words are separated by spaces), while leaving the first and last letter
the same.  If the last character in word in punctuation such as a comma,
period, question mark, et cetera, it will leave both that and the
second-to-last letter in place.